opex-deficiencies-report: move promptUser to a new prompts helper and rename it 'ask'
